Change level columns to decimals

+++ database/migrations/2018_05_07_162518_create_exercise_workout_pivot_table.php
@@ -19,7 +19,7 @@ class CreateExerciseWorkoutPivotTable extends Migration
             $table->integer('workout_id')->unsigned()->index();
             $table->integer('unit_id')->unsigned()->index();
 
-            $table->integer('level');
+            $table->decimal('level', 10, 1);
             $table->integer('quantity');
 
             $table->foreign('exercise_id')->references('id')->on('exercises')->onDelete('cascade');
